requirements.txt
python-telegram-bot==13.15
requests==2.31.0
pandas==2.0.3
matplotlib==3.8.0
numpy==1.26.4
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.1
pytz==2024.1